Small change to Animation class. (#6)
* Small change to Animation class.

Previously getFrame() would only increment framePointer once even if
the time since last getFrame() call was greater than say the next 2 frame delays.

This commit makes getFrame() able to skip over multiple frames if the
elapsed time is greater than the skipped frames total delay. Making it
more accurate.

* Spotted an error in my last commit.

I was resetting m_overlappedTime when i should be incrementing to it.

This way even if the totalElapsed time is less than the current frames
delay it will be added to m_overlappedTime and be taken in to
consideration next call.

* More changes to Animation class

* Removed unnecessary framepointer check

The if statement checking if the framepointer is out of
bound is not necessary if we instead just use the modulo operator.

Technically this may eventually result in the framepointer overflowing
but it will just overflow back to 0 so it shouldn't be a problem.